requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=tests,dev --output-file=requirements.txt pyproject.toml
#
appdirs==1.4.4
    # via fs
attrs==23.1.0
    # via
    #   fontmake
    #   ufolib2
axisregistry==0.4.3
    # via fontbakery
babelfont==3.0.1
    # via
    #   collidoscope
    #   glyphtools
beautifulsoup4==4.12.2
    # via fontbakery
beziers==0.5.0
    # via
    #   fontbakery
    #   fontfeatures
    #   glyphtools
black==23.7.0
    # via nupuram (pyproject.toml)
booleanoperations==0.9.0
    # via ufo2ft
brotli==1.0.9
    # via nupuram (pyproject.toml)
cachetools==5.3.1
    # via tox
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via cmarkgfm
cffsubr==0.2.9.post1
    # via ufo2ft
chardet==5.1.0
    # via tox
charset-normalizer==3.2.0
    # via requests
click==8.1.5
    # via black
cmarkgfm==2022.10.27
    # via fontbakery
collidoscope==0.6.5
    # via fontbakery
colorama==0.4.6
    # via tox
commandlines==0.4.1
    # via ufolint
compreffor==0.5.3
    # via ufo2ft
cu2qu==1.6.7.post2
    # via babelfont
defcon==0.10.2
    # via
    #   fontbakery
    #   glyphsets
dehinter==4.0.0
    # via fontbakery
distlib==0.3.6
    # via virtualenv
filelock==3.12.2
    # via
    #   tox
    #   virtualenv
font-v==2.1.0
    # via fontbakery
fontbakery[freetype]==0.8.13
    # via nupuram (pyproject.toml)
fontfeatures==1.7.4
    # via
    #   babelfont
    #   nupuram (pyproject.toml)
fontmake==3.6.0
    # via nupuram (pyproject.toml)
fontmath==0.9.3
    # via fontmake
fonttools[lxml,ufo,unicode]==4.41.0
    # via
    #   axisregistry
    #   babelfont
    #   booleanoperations
    #   cffsubr
    #   collidoscope
    #   compreffor
    #   cu2qu
    #   defcon
    #   dehinter
    #   font-v
    #   fontbakery
    #   fontfeatures
    #   fontmake
    #   fontmath
    #   glyphsets
    #   glyphslib
    #   kurbopy
    #   ufo2ft
    #   ufolib2
    #   ufolint
    #   vharfbuzz
freetype-py==2.3.0
    # via fontbakery
fs==2.4.16
    # via
    #   fontfeatures
    #   fonttools
gflanguages==0.8.9
    # via fontbakery
gitdb==4.0.10
    # via gitpython
gitpython==3.1.32
    # via font-v
glyphsets==0.6.2
    # via fontbakery
glyphslib==6.2.5
    # via
    #   babelfont
    #   fontmake
    #   glyphsets
    #   glyphtools
glyphtools==0.8.0
    # via fontfeatures
idna==3.4
    # via requests
isort==5.12.0
    # via nupuram (pyproject.toml)
kurbopy==0.9.0
    # via collidoscope
lxml==4.9.3
    # via
    #   fontbakery
    #   fontfeatures
    #   fonttools
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
munch==4.0.0
    # via nupuram (pyproject.toml)
munkres==1.1.4
    # via fontbakery
mypy-extensions==1.0.0
    # via black
numpy==1.25.1
    # via glyphtools
openstep-plist==0.3.0.post1
    # via
    #   babelfont
    #   glyphslib
opentype-sanitizer==9.1.0
    # via fontbakery
opentypespec==1.9.1
    # via fontbakery
orjson==3.9.2
    # via babelfont
packaging==23.1
    # via
    #   black
    #   fontbakery
    #   pyproject-api
    #   tox
pathspec==0.11.1
    # via black
pip-api==0.0.30
    # via fontbakery
platformdirs==3.8.1
    # via
    #   black
    #   tox
    #   virtualenv
pluggy==1.2.0
    # via tox
protobuf==3.20.3
    # via
    #   axisregistry
    #   fontbakery
    #   gflanguages
pyclipper==1.3.0.post4
    # via
    #   beziers
    #   booleanoperations
pycparser==2.21
    # via cffi
pygments==2.15.1
    # via rich
pyproject-api==1.5.3
    # via tox
pyyaml==6.0
    # via
    #   fontbakery
    #   nupuram (pyproject.toml)
requests==2.31.0
    # via fontbakery
rich==13.4.2
    # via fontbakery
rstr==3.2.1
    # via stringbrewer
ruff==0.0.278
    # via nupuram (pyproject.toml)
six==1.16.0
    # via fs
skia-pathops==0.8.0
    # via collidoscope
smmap==5.0.0
    # via gitdb
soupsieve==2.4.1
    # via beautifulsoup4
sre-yield==1.2
    # via stringbrewer
stringbrewer==0.0.1
    # via fontbakery
toml==0.10.2
    # via fontbakery
tox==4.6.4
    # via nupuram (pyproject.toml)
tqdm==4.65.0
    # via collidoscope
ufo2ft[compreffor]==2.32.0
    # via
    #   fontbakery
    #   fontmake
ufolib2==0.14.0
    # via
    #   babelfont
    #   fontmake
    #   glyphslib
ufolint==1.2.0
    # via fontbakery
ufonormalizer==0.6.1
    # via nupuram (pyproject.toml)
uharfbuzz==0.37.0
    # via
    #   collidoscope
    #   nupuram (pyproject.toml)
    #   vharfbuzz
unicodedata2==15.0.0
    # via
    #   fontbakery
    #   fonttools
urllib3==2.0.3
    # via requests
vharfbuzz==0.2.0
    # via fontbakery
virtualenv==20.23.1
    # via tox

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
